requirements.txt
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
Werkzeug==2.1.2
